+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/UTxO.hs
@@ -104,8 +104,7 @@ instance EraUTxO AlonzoEra where
 
   getConsumedValue = getConsumedMaryValue
 
-  getProducedValue pp isRegPoolId txBody =
-    withTopTxLevelOnly txBody (getProducedMaryValue pp isRegPoolId)
+  getProducedValue = getProducedMaryValue
 
   getScriptsProvided _ tx = ScriptsProvided (tx ^. witsTxL . scriptTxWitsL)